requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
Pygments==2.15.1
PyJWT==2.7.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
rich==13.4.2
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
